+++ database/buildSchema/06_trigger_queue.sql
@@ -1,6 +1,6 @@
 -- trigger queue
 
-CREATE TYPE public.trigger as ENUM ('onApplicationCreate', 'onApplicationSubmit', 'onApplicationSave', 'onApplicationWithdraw', 'onReviewCreate', 'onReviewSubmit', 'onReviewStart', 'onReviewAssign', 'onApprovalSubmit', 'onScheduleTime', 'Processing', 'Error');
+CREATE TYPE public.trigger as ENUM ('onApplicationCreate', 'onApplicationSubmit', 'onApplicationSave', 'onApplicationWithdraw', 'onReviewCreate', 'onReviewSubmit', 'onReviewStart', 'onReviewAssign', 'onReviewSelfAssign', 'onApprovalSubmit', 'onScheduleTime', 'Processing', 'Error');
 
 CREATE TYPE public.trigger_queue_status as ENUM ('Triggered', 'Actions Dispatched', 'Error');
