+++ database/insertData/03_template_D_basicReviewTesting.js
@@ -330,12 +330,16 @@ exports.queries = [
               }
               {
                 actionCode: "updateReviewAssignmentsStatus"
-                trigger: ON_REVIEW_ASSIGN
+                trigger: ON_REVIEW_SELF_ASSIGN
                 # sequence: 1
                 parameterQueries: {
                   reviewAssignmentId: {
                     operator: "objectProperties"
                     children: ["applicationData.record_id"]
+                  }
+                  trigger: {
+                    operator: "objectProperties"
+                    children: ["applicationData.trigger"]
                   }
                 }
               }
