+++ src/components/postgresConnect.ts
@@ -777,73 +777,6 @@ class PostgresDB {
     return reviewAssignmentIds
   }
 
-  public getReviewAssignmentById = async (reviewAssignmentId: number) => {
-    const text = `
-    SELECT * FROM review_assignment
-    WHERE id = $1
-    `
-    try {
-      const result = await this.query({
-        text,
-        values: [reviewAssignmentId],
-      })
-      return result.rows[0]
-    } catch (err) {
-      console.log(err.message)
-      throw err
-    }
-  }
-
-  public getMatchingReviewAssignments = async (
-    reviewAssignmentId: number,
-    applicationId: number,
-    stageNumber: number,
-    reviewLevel: number
-  ) => {
-    const text = `
-    SELECT * FROM review_assignment
-    WHERE application_id = $2
-    AND stage_number = $3
-    AND level = $4
-    AND id <> $1
-    `
-    try {
-      const result = await this.query({
-        text,
-        values: [reviewAssignmentId, applicationId, stageNumber, reviewLevel],
-      })
-      return result.rows
-    } catch (err) {
-      console.log(err.message)
-      throw err
-    }
-  }
-
-  public updateReviewAssignments = async (reviewAssignments: any) => {
-    const reviewAssignmentUpdateResults = []
-    for (const reviewAssignment of reviewAssignments) {
-      const { id, status } = reviewAssignment
-      const text = `
-      UPDATE review_assignment
-      SET status = $2
-      WHERE id = $1
-      RETURNING id, status
-      `
-      try {
-        const result = await this.query({
-          text,
-          values: [id, status],
-        })
-        reviewAssignmentUpdateResults.push(result.rows[0])
-      } catch (err) {
-        console.log(err.message)
-        reviewAssignmentUpdateResults.push(err.message)
-        throw err
-      }
-    }
-    return reviewAssignmentUpdateResults
-  }
-
   public isFullyAssignedLevel1 = async (applicationId: number) => {
     const text = `
     SELECT is_fully_assigned_level_1
