+++ src/plugins/action_update_review_assignments/src/types.ts
@@ -1,6 +1,6 @@
 export enum AssignmentStatus {
   AVAILABLE = 'Available',
-  NOT_AVAILABLE = 'Not available',
+  SELF_ASSIGNED_OTHER = 'Self-assigned by another',
   ASSIGNED = 'Assigned',
   SELF_ASSIGN = 'Available for self-assignment',
 }
